relocate to more generic group by
