+++ core/src/main/java/module-info.java
@@ -1,3 +1,5 @@
+import io.questdb.griffin.engine.functions.groupby.RegressionSlopeFunctionFactory;
+
 /*******************************************************************************
  *     ___                  _   ____  ____
  *    / _ \ _   _  ___  ___| |_|  _ \| __ )
@@ -127,7 +129,7 @@ open module io.questdb {
             io.questdb.griffin.engine.functions.finance.MidPriceFunctionFactory,
             io.questdb.griffin.engine.functions.finance.WeightedMidPriceFunctionFactory,
             io.questdb.griffin.engine.functions.finance.SpreadBpsFunctionFactory,
-            io.questdb.griffin.engine.functions.finance.RegressionSlopeFunctionFactory,
+            RegressionSlopeFunctionFactory,
 
 
             // query activity functions
