+++ core/src/main/resources/META-INF/services/io.questdb.griffin.FunctionFactory
@@ -27,7 +27,6 @@ io.questdb.griffin.engine.functions.finance.SpreadFunctionFactory
 io.questdb.griffin.engine.functions.finance.MidPriceFunctionFactory
 io.questdb.griffin.engine.functions.finance.WeightedMidPriceFunctionFactory
 io.questdb.griffin.engine.functions.finance.SpreadBpsFunctionFactory
-io.questdb.griffin.engine.functions.finance.RegressionSlopeFunctionFactory
 
 io.questdb.griffin.engine.functions.activity.CancelQueryFunctionFactory
 io.questdb.griffin.engine.functions.activity.QueryActivityFunctionFactory
@@ -716,6 +715,8 @@ io.questdb.griffin.engine.functions.groupby.ApproxCountDistinctLongGroupByDefaul
 io.questdb.griffin.engine.functions.groupby.ApproxCountDistinctLongGroupByFunctionFactory
 io.questdb.griffin.engine.functions.groupby.ApproxCountDistinctIPv4GroupByDefaultFunctionFactory
 io.questdb.griffin.engine.functions.groupby.ApproxCountDistinctIPv4GroupByFunctionFactory
+
+io.questdb.griffin.engine.functions.groupby.RegressionSlopeFunctionFactory
 
 # 'isOrdered'
 io.questdb.griffin.engine.functions.groupby.IsIPv4OrderedGroupByFunctionFactory
